% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{add_active_site_indexes}
\alias{add_active_site_indexes}
\title{add active-site indexes to prot object}
\usage{
add_active_site_indexes(prot, pdb_site_active)
}
\description{
add active-site indexes to prot object
}
