% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_enm.R
\name{add_enm}
\alias{add_enm}
\title{Add \code{enm} object to \code{prot} object}
\usage{
add_enm(prot, model, d_max, frustrated, ...)
}
\arguments{
\item{prot}{A protein object that must contain \code{xyz} and \code{pdb_site}}

\item{model}{the enm model}

\item{d_max}{a cdistance cut-off needed by some models to define sites in contact}
}
\value{
A protein object equal to input with enm added, where enm is a list containing \code{graph, eij, kmat, mode, evalue, cmat, umat}
}
\description{
Add \code{enm} object to \code{prot} object
}
\seealso{
Other enm builders: \code{\link{add_enm_activation}},
  \code{\link{enm_energy_activation}},
  \code{\link{enm_energy}}, \code{\link{enm_graph_xyz}},
  \code{\link{enm_nma_basis}}, \code{\link{enm_nma}},
  \code{\link{enm_set_xyz}}, \code{\link{kmat_graph}}
}
\concept{enm builders}
