% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{add_enm_activation}
\alias{add_enm_activation}
\title{Add cmat_activation and kmat_activation to prot object}
\usage{
add_enm_activation(prot)
}
\arguments{
\item{prot}{A protein object that must contain \code{xyz} and \code{pdb_site} elements, and active_site indexes \code{ind_active} (which may be NA)}
}
\value{
A protein object with added cmat_active and kmat_active
}
\description{
Add cmat_activation and kmat_activation to prot object
}
\seealso{
Other enm builders: 
\code{\link{enm_nma_basis}()},
\code{\link{enm_nma}()},
\code{\link{kmat_graph}()},
\code{\link{set_enm_graph}()}
}
\concept{enm builders}
