% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_admrs.R
\name{admrs}
\alias{admrs}
\title{Calculate double-mutational-scan matrix using analytical  method}
\usage{
admrs(wt, mut_dl_sigma, mut_sd_min, option = "mean_max")
}
\description{
Calculate double-mutational-scan matrix using analytical  method
}
