% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admrs.R
\name{admrs_amat}
\alias{admrs_amat}
\title{Calculate site-by-site analytic response matrix, general, o
minimize over mutations at i and 3yyj}
\usage{
admrs_amat(wt, amat, mut_dl_sigma, mut_sd_min, option)
}
\description{
Calculate site-by-site analytic response matrix, general, o
minimize over mutations at i and 3yyj
}
