% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_amrs.R
\name{amrs_all}
\alias{amrs_all}
\title{Calculate all response matrices and profiles, "fast prs" method}
\usage{
amrs_all(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate all response matrices and profiles, "fast prs" method
}
