% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy.R
\name{beta_boltzmann}
\alias{beta_boltzmann}
\title{Boltzmann's beta, \eqn{1 / RT}}
\usage{
beta_boltzmann(R = 0.001986, T = 298)
}
\arguments{
\item{R}{Boltzmann's constant per mole, in kcal/(mol K)}

\item{T}{absolute temperature, in Kelvin}
}
\value{
a scalar, \eqn{1 / (RT)}, in mol/kcal
}
\description{
The default inverse temperature used by the entropic energy functions
([ddg_tds()], [ddgact_tds()], [dgact_tds()]).
}
\examples{
beta_boltzmann()              # 298 K
beta_boltzmann(T = 310)       # body temperature

}
\seealso{
Other enm_energy:
\code{\link[=dgact_dv]{dgact_dv()}},
\code{\link[=dgact_tds]{dgact_tds()}},
\code{\link[=enm_g_entropy]{enm_g_entropy()}},
\code{\link[=enm_v_min]{enm_v_min()}}
}
\concept{enm_energy}
