% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{calculate_de2ij.fast}
\alias{calculate_de2ij.fast}
\title{Calculate site-by-site energy response matrix de2(n, j) using method "fast"}
\usage{
calculate_de2ij.fast(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate site-by-site energy response matrix de2(n, j) using method "fast"
}
