% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{calculate_de2ij.new}
\alias{calculate_de2ij.new}
\title{Calculate energy response matrix de2(i, j) using method "new"}
\usage{
calculate_de2ij.new(wt, fmat)
}
\description{
Calculate energy response matrix de2(i, j) using method "new"
}
