% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_mrs.R
\name{calculate_de2ij.prs}
\alias{calculate_de2ij.prs}
\title{Calculate site-by-site  energy response matrix de2(i, j) using method "sim"}
\usage{
calculate_de2ij.prs(mutants)
}
\description{
Calculate site-by-site  energy response matrix de2(i, j) using method "sim"
}
