% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_smrs.R
\name{calculate_de2ij_smrs}
\alias{calculate_de2ij_smrs}
\title{Calculate energy response matrix de2(i, j) using method "new"}
\usage{
calculate_de2ij_smrs(wt, fmat)
}
\description{
Calculate energy response matrix de2(i, j) using method "new"
}
