% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{calculate_de2j.prs}
\alias{calculate_de2j.prs}
\title{calculate influence profile de2j (relaxation)}
\usage{
calculate_de2j.prs(mutants)
}
\arguments{
\item{mutants}{is a set of mutants generated by get_mutants_table}
}
\value{
an influence vector, where each element is the total response to a mutation at site j
}
\description{
calculate influence profile de2j (relaxation)
}
