% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{calculate_de2nj.new}
\alias{calculate_de2nj.new}
\title{Calculate mode-by-site energy response matrix de2(n, j) using method "new"}
\usage{
calculate_de2nj.new(wt, fmat)
}
\description{
Calculate mode-by-site energy response matrix de2(n, j) using method "new"
}
