% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{calculate_de2nj.prs}
\alias{calculate_de2nj.prs}
\title{Calculate mode-by-site energy response matrix de2(n, j) using method "sim"}
\usage{
calculate_de2nj.prs(mutants)
}
\description{
Calculate mode-by-site energy response matrix de2(n, j) using method "sim"
}
