% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_amrs.R
\name{calculate_de2nj_amrs}
\alias{calculate_de2nj_amrs}
\title{Calculate mode-by-site energy response matrix de2(n, j) using method "fast"}
\usage{
calculate_de2nj_amrs(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate mode-by-site energy response matrix de2(n, j) using method "fast"
}
