% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{calculate_df2ij.new}
\alias{calculate_df2ij.new}
\title{Calculate force response matrix df2(i, j) using method "new"}
\usage{
calculate_df2ij.new(wt, fmat)
}
\description{
Calculate force response matrix df2(i, j) using method "new"
}
