% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_smrs.R
\name{calculate_df2ij_smrs}
\alias{calculate_df2ij_smrs}
\title{Calculate force response matrix df2(i, j) using method "new"}
\usage{
calculate_df2ij_smrs(wt, fmat)
}
\description{
Calculate force response matrix df2(i, j) using method "new"
}
