% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{calculate_df2nj.fast}
\alias{calculate_df2nj.fast}
\title{Calculate mode-by-site force matrix df2(n, j) using method "fast"}
\usage{
calculate_df2nj.fast(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate mode-by-site force matrix df2(n, j) using method "fast"
}
