% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_smrs.R
\name{calculate_df2nj_smrs}
\alias{calculate_df2nj_smrs}
\title{Calculate mode-by-site force response matrix df2(n, j) using method "new"}
\usage{
calculate_df2nj_smrs(wt, fmat)
}
\description{
Calculate mode-by-site force response matrix df2(n, j) using method "new"
}
