% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{calculate_dfj.fast}
\alias{calculate_dfj.fast}
\title{Calculate several influence profiles}
\usage{
calculate_dfj.fast(wt, mut_dl_sigma, mut_sd_min, ...)
}
\description{
Calculate several influence profiles
}
