% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{calculate_dfnj.fast}
\alias{calculate_dfnj.fast}
\title{Calculate structural mutational response, mode analysis}
\usage{
calculate_dfnj.fast(wt, mut_dl_sigma = 0.3, mut_sd_min = 2)
}
\description{
Calculate structural mutational response, mode analysis
}
