% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{calculate_dfnj.prs}
\alias{calculate_dfnj.prs}
\title{Calculate structural mutational response, mode analysis}
\usage{
calculate_dfnj.prs(mutants)
}
\description{
Calculate structural mutational response, mode analysis
}
