% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{calculate_dr2i}
\alias{calculate_dr2i}
\alias{calculate_de2i}
\alias{calculate_df2i}
\title{Compare the structures of two proteins in site representation}
\usage{
calculate_dr2i(prot1, prot2)

calculate_de2i(prot1, prot2, kmat_sqrt)

calculate_df2i(prot1, prot2)
}
\arguments{
\item{prot1}{A protein object with \code{xyz} defined}

\item{prot2}{A second protein object  with \code{xyz} defined}
}
\value{
A tibble with columns \code{pdb_site, site, dr2}
}
\description{
Calculate de difference between the structures of two proteins, return dr2i
}
\details{
This version works only for prot1 and prot2 with no indels
}
