% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{calculate_dr2ij.new}
\alias{calculate_dr2ij.new}
\title{Calculate structure response matrix dr2(i, j) using method "new"}
\usage{
calculate_dr2ij.new(wt, fmat)
}
\description{
Calculate structure response matrix dr2(i, j) using method "new"
}
