% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{calculate_dr2ij.prs}
\alias{calculate_dr2ij.prs}
\title{Calculate site-by-site structure response matrix dr2(i, j) using method "sim"}
\usage{
calculate_dr2ij.prs(mutants)
}
\description{
Calculate site-by-site structure response matrix dr2(i, j) using method "sim"
}
