% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_smrs.R
\name{calculate_dr2ij_smrs}
\alias{calculate_dr2ij_smrs}
\title{Calculate structure response matrix dr2(i, j) using method "new"}
\usage{
calculate_dr2ij_smrs(wt, fmat)
}
\description{
Calculate structure response matrix dr2(i, j) using method "new"
}
