% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{calculate_dr2j.prs}
\alias{calculate_dr2j.prs}
\title{calculate influence profile dr2j (structure)}
\usage{
calculate_dr2j.prs(mutants)
}
\arguments{
\item{mutants}{is a set of mutants generated by get_mutants_table}
}
\value{
an influence vector, where each element is the total response to a mutation at site j
}
\description{
calculate influence profile dr2j (structure)
}
