% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_sim.R
\name{calculate_dr2nj.prs}
\alias{calculate_dr2nj.prs}
\title{Calculate mode-by-site structure response matrix dr2(n, j) using method "sim"}
\usage{
calculate_dr2nj.prs(mutants)
}
\description{
Calculate mode-by-site structure response matrix dr2(n, j) using method "sim"
}
