% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{calculate_dvmj.prs}
\alias{calculate_dvmj.prs}
\title{calculate influence profile dvmj (minimumn energy)}
\usage{
calculate_dvmj.prs(mutants)
}
\arguments{
\item{mutants}{is a set of mutants generated by get_mutants_table}
}
\value{
an influence vector, where each element is the total response to a mutation at site j
}
\description{
calculate influence profile dvmj (minimumn energy)
}
