% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{calculate_dvsi.noindel}
\alias{calculate_dvsi.noindel}
\title{Calculate dvsi, site-dependent stress energy}
\usage{
calculate_dvsi.noindel(wt, mut)
}
\description{
Calculate dvsi, site-dependent stress energy
}
