% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{calculate_dvsij.fast}
\alias{calculate_dvsij.fast}
\title{Calculate site-by-site stress-energy response matrix dvs(n, j) using method "fast"}
\usage{
calculate_dvsij.fast(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate site-by-site stress-energy response matrix dvs(n, j) using method "fast"
}
