% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{calculate_dvsij.new}
\alias{calculate_dvsij.new}
\title{Calculate stress energy response matrix dvs(i, j) using method "new"}
\usage{
calculate_dvsij.new(wt, dlmat)
}
\description{
Calculate stress energy response matrix dvs(i, j) using method "new"
}
