% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_amrs.R
\name{calculate_dvsij_amrs}
\alias{calculate_dvsij_amrs}
\title{Calculate site-by-site stress-energy response matrix dvs(n, j) using method "fast"}
\usage{
calculate_dvsij_amrs(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate site-by-site stress-energy response matrix dvs(n, j) using method "fast"
}
