% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_smrs.R
\name{calculate_dvsij_smrs}
\alias{calculate_dvsij_smrs}
\title{Calculate stress energy response matrix dvs(i, j) using method "new"}
\usage{
calculate_dvsij_smrs(wt, dlmat)
}
\description{
Calculate stress energy response matrix dvs(i, j) using method "new"
}
