% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{calculate_dvsj.prs}
\alias{calculate_dvsj.prs}
\title{calculate influence profile dvsj (stress)}
\usage{
calculate_dvsj.prs(mutants)
}
\arguments{
\item{mutants}{is a set of mutants generated by get_mutants_table}
}
\value{
an influence vector, where each element is the total response to a mutation at site j
}
\description{
calculate influence profile dvsj (stress)
}
