% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{calculate_enm_eij}
\alias{calculate_enm_eij}
\title{Calculate unit vectors of edges}
\usage{
calculate_enm_eij(xyz, i, j)
}
\arguments{
\item{xyz}{vector of xyz coordinates}

\item{i, j}{integer vectors of nodes connected in each edge}
}
\value{
matrix with n_edge rows and 3 columns (x, y, z)
}
\description{
Calculate unit vectors of edges
}
\seealso{
Other enm builders: 
\code{\link{calculate_enm_graph}()},
\code{\link{calculate_enm_kmat}()},
\code{\link{calculate_enm_nma}()},
\code{\link{calculate_enm_nodes}()}
}
\concept{enm builders}
\keyword{internal}
