% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{calculate_enm_kmat}
\alias{calculate_enm_kmat}
\title{Calculate kmat given the ENM graph}
\usage{
calculate_enm_kmat(graph, eij, nsites, frustrated)
}
\arguments{
\item{graph}{A tibble representing the ENM graph (with edge information, especially \code{kij}}

\item{eij}{A matrix of size \code{n_edges x 3} of \code{eij} versors directed along ENM contacts}

\item{nsites}{The number of nodes of the ENM network}

\item{frustrated}{Logical indicating  wether to add frustration or not before calculating \code{kmat}}
}
\value{
The \code{3 nsites x 3 nsites} stiffness matrix of the ENM
}
\description{
Calculate kmat given the ENM graph
}
\examples{
\dontrun{
pdb <- read.pdb("2acy")
nodes <- calculate_enm_nodes <- function(pdb, node = "ca")
graph <- calculate_enm_graph(nodes$xyz, nodes$pdb_site, model = "anm", d_max = 10.5)
eij <- calculate_enm_eij(nodes$xyz, graph$i, graph$j)
kmat <- calculate_enm_kmat(graph, eij, nsite = )
}

}
\seealso{
Other enm builders: 
\code{\link{calculate_enm_eij}()},
\code{\link{calculate_enm_graph}()},
\code{\link{calculate_enm_nma}()},
\code{\link{calculate_enm_nodes}()}
}
\concept{enm builders}
