% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{calculate_fast_response_matrix_mode}
\alias{calculate_fast_response_matrix_mode}
\title{Calculate fast response matrix}
\usage{
calculate_fast_response_matrix_mode(wt, avector, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate fast response matrix
}
