% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_sdmrs.R
\name{calculate_force_dmrs}
\alias{calculate_force_dmrs}
\title{get force from delta_lik for dmrs calculations}
\usage{
calculate_force_dmrs(wt, delta_lij)
}
\description{
get force from delta_lik for dmrs calculations
}
