% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_amrs.R
\name{calculate_Rij_amrs}
\alias{calculate_Rij_amrs}
\title{Calculate site-by-site "fast" response matrix, general}
\usage{
calculate_Rij_amrs(wt, amat, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate site-by-site "fast" response matrix, general
}
