% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_amrs.R
\name{calculate_Rnj_amrs}
\alias{calculate_Rnj_amrs}
\title{Calculate fast response matrix, general}
\usage{
calculate_Rnj_amrs(wt, avector, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate fast response matrix, general
}
