% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cn_xyz}
\alias{cn_xyz}
\title{Title Calculate contact number of sites}
\usage{
cn_xyz(xyz, d_max)
}
\arguments{
\item{xyz}{xyz coordinates of sites (e.g. alpha carbons, or center of mass, etc.)}

\item{d_max}{cutoff radius to define contacts}
}
\value{
vector of contact numbers
}
\description{
Calculates number of contacts.
}
