% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{delta_energy_site}
\alias{delta_energy_site}
\title{Calculate structural mutational response, site analysis}
\usage{
delta_energy_site(mutants)
}
\description{
Calculate structural mutational response, site analysis
}
