% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutant_scan_functions.R
\name{delta_structure_mode}
\alias{delta_structure_mode}
\title{Calculate structural mutational response, mode analysis}
\usage{
delta_structure_mode(mutants)
}
\description{
Calculate structural mutational response, mode analysis
}
