% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutant_scan_functions.R
\name{delta_structure_site}
\alias{delta_structure_site}
\title{Calculate structural mutational response, site analysis}
\usage{
delta_structure_site(mutants)
}
\description{
Calculate structural mutational response, site analysis
}
