% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{delta_v_stress}
\alias{delta_v_stress}
\title{Stress-model difference of local-mutational-stress energy}
\usage{
delta_v_stress(prot1, prot2, ideal = prot1)
}
\description{
Stress-model difference of local-mutational-stress energy
}
