% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_fast.R
\name{dmrs_analytical}
\alias{dmrs_analytical}
\title{Calculate double-mutational-scan matrix using analytical "fast" method}
\usage{
dmrs_analytical(wt, mut_dl_sigma, mut_sd_min, option = "mean_min")
}
\description{
Calculate double-mutational-scan matrix using analytical "fast" method
}
