% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_fast.R
\name{dmrs_analytical_amat}
\alias{dmrs_analytical_amat}
\title{Calculate site-by-site "fast" response matrix, general, o
minimize over mutations at i and 3yyj}
\usage{
dmrs_analytical_amat(wt, amat, mut_dl_sigma, mut_sd_min, option)
}
\description{
Calculate site-by-site "fast" response matrix, general, o
minimize over mutations at i and 3yyj
}
