% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms.R
\name{dmrs_simulation}
\alias{dmrs_simulation}
\title{Calculate double-mutational-scanning matrix}
\usage{
dmrs_simulation(wt, nmut, mut_dl_sigma, mut_sd_min, seed, option = "mean_min")
}
\description{
Calculate double-mutational-scanning matrix
}
