% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms.R
\name{dmsmat_dms.fast}
\alias{dmsmat_dms.fast}
\title{Calculate double-mutational-scan matrix using analytical "fast" method}
\usage{
dmsmat_dms.fast(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate site-by-site structure response matrix dr2(n, j) using method "fast"
}
