% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms.R
\name{dmsmat_dms.new}
\alias{dmsmat_dms.new}
\title{Calculate double-mutational-scanning matrix}
\usage{
dmsmat_dms.new(wt, nmut, mut_dl_sigma, mut_sd_min, seed)
}
\description{
Calculate double-mutational-scanning matrix
}
