% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_new_alternatives.R
\name{dmsmat_dms.new_v2}
\alias{dmsmat_dms.new_v2}
\title{Calculate double-mutational-scanning matrix}
\usage{
dmsmat_dms.new_v2(wt, nmut, mut_dl_sigma, mut_sd_min, seed)
}
\description{
Calculate double-mutational-scanning matrix
}
