% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfenm_functions.R
\name{dr2_nm}
\alias{dr2_nm}
\alias{de2_nm}
\alias{df2_nm}
\title{Compare the structures of two proteins in nm representation}
\usage{
dr2_nm(prot1, prot2)

de2_nm(prot1, prot2)

df2_nm(prot1, prot2)
}
\arguments{
\item{prot1}{A protein object with \code{xyz} and \code{enm} defined}

\item{prot2}{A second protein object  with \code{xyz} defined}
}
\value{
A vector of square differences along normal modes \code{dr2n}
}
\description{
Calculate de difference between the structures of two proteins, return dr2_nm.
}
\details{
This version works only for prot1 and prot2 with no indels
}
