% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{dr2_site}
\alias{dr2_site}
\alias{de2_site}
\alias{df2_site}
\title{Compare the structures of two proteins in site representation}
\usage{
dr2_site(prot1, prot2)

de2_site(prot1, prot2)

df2_site(prot1, prot2)
}
\arguments{
\item{prot1}{A protein object with \code{xyz} defined}

\item{prot2}{A second protein object  with \code{xyz} defined}
}
\value{
A tibble with columns \code{pdb_site, site, dr2}
}
\description{
Calculate de difference between the structures of two proteins, return dr2_site.
}
\details{
This version works only for prot1 and prot2 with no indels
}
