% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_add.R
\name{eij_edge}
\alias{eij_edge}
\title{Calculate unit vectors of edges}
\usage{
eij_edge(xyz, i, j)
}
\arguments{
\item{xyz}{vector of xyz coordinates}

\item{i, j}{integer vectors of nodes connected in each edge}
}
\value{
matrix with n_edge rows and 3 columns (x, y, z)
}
\description{
Calculate unit vectors of edges
}
