% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_add.R
\name{enm_add}
\alias{enm_add}
\alias{add_enm}
\title{Add \code{enm} object to \code{prot} object}
\usage{
enm_add(prot, param_list)

add_enm(prot, param_list)
}
\arguments{
\item{prot}{A protein object that must contain \code{xyz} and \code{pdb_site} elements, and active_site indexes \code{ind_active} (which may be NA)}

\item{param_list}{A list of \code{enm} parameters including elements \code{model, k_scale, d_max, sd_min}}
}
\value{
A protein object equal to input with enm added, where enm is a list containing \code{graph, eij, kmat, mode, evalue, cmat, umat}
 and also \code{cmat_active, kmat_active}, which are NA if \code{ind_active} is NA.
}
\description{
Add \code{enm} object to \code{prot} object
}
\seealso{
Other enm builders: \code{\link{enm_energy}},
  \code{\link{enm_graph_xyz}}, \code{\link{enm_nma_basis}},
  \code{\link{enm_nma}}, \code{\link{enm_set_xyz}},
  \code{\link{enm_set}}, \code{\link{enm_v0_add}},
  \code{\link{kmat_graph}}
}
\concept{enm builders}
