% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy.R
\name{enm_dv_activation}
\alias{enm_dv_activation}
\title{Internal energy contribution to free energy of activaton}
\usage{
enm_dv_activation(prot, ideal)
}
\description{
Internal energy contribution to free energy of activaton
}
