% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy.R
\name{enm_energy}
\alias{enm_energy}
\alias{energy}
\title{Calculate various enm energies for prot object}
\usage{
enm_energy(prot, ideal, sd_min = 1, beta = beta_boltzmann())

energy(prot, ideal, sd_min = 1, beta = beta_boltzmann())
}
\arguments{
\item{prot}{A protein object, with enm graph defined}

\item{ideal}{A protein object that corresponds to the ideal "active" conformation}

\item{sd_min}{Integer representing a cut-off in sequence distance for energy calculations}

\item{beta}{boltzmann temperature}
}
\value{
A list of energy terms \code{lst(v_min, dv_activation, g_entropy, g_entropy_activation, v_stress)}
}
\description{
Given a protein (prot) with enm defined, it calculates energy terms
returns a list of internal energies (v) and entropic terms (g)
(note: hemholtz free energy is A = v_min + g_energy)
}
\examples{

}
\seealso{
Other enm builders: \code{\link{enm_add}},
  \code{\link{enm_graph_xyz}}, \code{\link{enm_nma_basis}},
  \code{\link{enm_nma}}, \code{\link{enm_set_xyz}},
  \code{\link{kmat_graph}}
}
\concept{enm builders}
