% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{enm_from_xyz}
\alias{enm_from_xyz}
\title{Set up ENM model}
\usage{
enm_from_xyz(xyz, pdb_site, model, d_max, frustrated, ...)
}
\arguments{
\item{xyz}{The \code{3 x N} matrix containing the Cartesian coordinates of ENM nodes}

\item{pdb_site}{The pdb numbering of network nodes}

\item{model}{The ENM model variant (gnm, anm, hnm0, hnm, ming_wall, reach, pfgnm)}

\item{d_max}{The cut-off used to define contacts in some models}

\item{...}{Any other parameter}
}
\value{
A list containing the ENM graph `graph`, the matrix of versors `eij`, and the ENM K matrix `kmat`
}
\description{
Set up ENM model from node coordinates `xyz` and residue pdb numbers `pdb_site`.
}
\examples{

}
\seealso{
Other enm builders: 
\code{\link{add_enm_activation}()},
\code{\link{enm_graph_xyz}()},
\code{\link{enm_nma_basis}()},
\code{\link{enm_nma}()},
\code{\link{kmat_graph}()}
}
\concept{enm builders}
