% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{enm_g_entropy_activation}
\alias{enm_g_entropy_activation}
\title{entropic contribution to dg_activation}
\usage{
enm_g_entropy_activation(prot, beta)
}
\description{
entropic contribution to dg_activation
}
