% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{enm_graph_xyz}
\alias{enm_graph_xyz}
\title{Calculate ENM graph}
\usage{
enm_graph_xyz(xyz, pdb_site, model, d_max, ...)
}
\arguments{
\item{xyz}{matrix of size \code{c(3,N)} containing each column the \code{x, y, z} coordinates of each of N nodes}

\item{pdb_site}{integer vector of size N containing the number of each node (pdb residue number)}

\item{model}{character variable specifying the ENM model variant, default is \code{"gnm"}, options:
\code{gnm, anm, ming_wall, hnm, hnm0, pfgnm, reach}.}

\item{d_max}{distance-cutoff to define network contacts}
}
\value{
a tibble that contains the graph representation of the network
}
\description{
Calculates graph representation of Elastic Network Model (ENM), the typical relaxed case (lij = dij)
}
\examples{
enm_graph_xyz(xyz, pdb_site)

}
\seealso{
Other enm builders: 
\code{\link{add_enm_activation}()},
\code{\link{enm_nma_basis}()},
\code{\link{enm_nma}()},
\code{\link{kmat_graph}()}
}
\concept{enm builders}
