% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_nma.R
\name{enm_nma}
\alias{enm_nma}
\title{Perform Normal Mode Analysis}
\usage{
enm_nma(kmat, TOL = 1e-05)
}
\arguments{
\item{kmat}{The K matrix to diagonalize}

\item{TOL=1.e-5}{A small value, eigenvectors with eigenvalues larger than `TOL` are discarded}
}
\value{
A list with elements \code{mode}, \code{evalue}, \code{cmat}, and \code{umat}
}
\description{
Given an enm `kmat`, perform NMA
}
\examples{

enm_anm(enm, 1.e-10)

}
\seealso{
Other enm builders: 
\code{\link{add_enm_activation}()},
\code{\link{enm_from_xyz}()},
\code{\link{enm_graph_xyz}()},
\code{\link{enm_nma_basis}()},
\code{\link{kmat_graph}()}
}
\concept{enm builders}
