% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfenm_functions.R
\name{enm_update}
\alias{enm_update}
\alias{update_enm}
\title{Update ENM following graph change}
\usage{
enm_update(prot, add_frust = FALSE)

update_enm(prot, add_frust = FALSE)
}
\arguments{
\item{prot}{A protein object}

\item{add_frust}{A logical indicating whether to add frustration}
}
\value{
A protein object with updated \code{enm} component
}
\description{
Update ENM following graph change
}
\examples{

}
\seealso{
Other enm mutating functions: \code{\link{get_force}},
  \code{\link{get_mutant_site_2}},
  \code{\link{get_mutant_site}}, \code{\link{get_mutant}}
}
\concept{enm mutating functions}
