% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_add.R
\name{enm_v0_add}
\alias{enm_v0_add}
\alias{add_v0}
\title{Add v0 to ENM}
\usage{
enm_v0_add(prot, v0 = 0)

add_v0(prot, v0 = 0)
}
\arguments{
\item{prot}{A protein object}

\item{v0}{A scalar, representing the ENM energy at the minimum}
}
\value{
A protein object identical to input, but with v0 added
}
\description{
Add \code{v0ij = v0 / nedges} to each edge of graph of enm object
}
\examples{

}
\seealso{
Other enm builders: \code{\link{enm_add}},
  \code{\link{enm_energy}}, \code{\link{enm_graph_xyz}},
  \code{\link{enm_nma_basis}}, \code{\link{enm_nma}},
  \code{\link{enm_set_xyz}}, \code{\link{enm_set}},
  \code{\link{kmat_graph}}
}
\concept{enm builders}
