% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{fast_calculate_dfij}
\alias{fast_calculate_dfij}
\title{Calculate structural mutational response, site analysis}
\usage{
fast_calculate_dfij(wt, mut_dl_sigma = 0.3, mut_sd_min = 2)
}
\description{
Calculate structural mutational response, site analysis
}
