% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{fast_calculate_dfj}
\alias{fast_calculate_dfj}
\title{Calculate energy mutational response}
\usage{
fast_calculate_dfj(wt, mut_dl_sigma, mut_sd_min, ...)
}
\description{
Calculate energy mutational response
}
