% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{fast_calculate_resp_nj}
\alias{fast_calculate_resp_nj}
\title{Calculate fast response matrix}
\usage{
fast_calculate_resp_nj(wt, avector, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate fast response matrix
}
