% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{fast_delta_energy}
\alias{fast_delta_energy}
\title{Calculate energy mutational response}
\usage{
fast_delta_energy(wt, mut_dl_sigma, mut_sd_min, ...)
}
\description{
Calculate energy mutational response
}
