% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{fast_delta_structure_mode}
\alias{fast_delta_structure_mode}
\title{Calculate structural mutational response, mode analysis}
\usage{
fast_delta_structure_mode(wt, mut_dl_sigma = 0.3, mut_sd_min = 2)
}
\description{
Calculate structural mutational response, mode analysis
}
