% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmrs.R
\name{generate_fmat_dmrs}
\alias{generate_fmat_dmrs}
\title{Generate force-matrix for dmrs_matrix calculation}
\usage{
generate_fmat_dmrs(wt, nmut, mut_dl_sigma, mut_sd_min, seed)
}
\description{
Generate force-matrix for dmrs_matrix calculation
}
