% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{generate_perturbations}
\alias{generate_perturbations}
\title{Generate matrices of perturbations (dlmat) and forces (fmat)}
\usage{
generate_perturbations(wt, nmut, mut_dl_sigma, mut_sd_min, seed)
}
\description{
dlmat(edge, j, m) is the perturbation dl at edge e due to mutation m at site j
fmat(i, j, m) is force at site i for mutation m = c(1,...nmut), at site j
}
