% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_bfactor}
\alias{get_bfactor}
\title{Get the X-ray b-factor}
\usage{
get_bfactor(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
a vector of X-ray B-factors for ENM nodes
}
\description{
Get the X-ray b-factor
}
