% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_cmat}
\alias{get_cmat}
\title{Get ENM covariance matrix}
\usage{
get_cmat(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
cmat, the covariance matrix of the ENM network, of size 3 nsites x 3 nsites
}
\description{
Get ENM covariance matrix
}
