% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_d_max}
\alias{get_d_max}
\title{Get ENM contact distance cut-off}
\usage{
get_d_max(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
d_max, the cut-off used to build the ENM network
}
\description{
Get ENM contact distance cut-off
}
