% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_eij}
\alias{get_eij}
\title{Get the unit vectors eij}
\usage{
get_eij(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
a matrix of size nedges x 3, containing unit vectors eij for all edges
}
\description{
Get the unit vectors eij
}
