% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_enm_model}
\alias{get_enm_model}
\title{Get ENM model type}
\usage{
get_enm_model(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
ENM model type ("anm", "hnm0", "hnm", "ming_wall", "pfanm")
}
\description{
Get ENM model type
}
