% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_enm_node}
\alias{get_enm_node}
\title{Get ENM node-type}
\usage{
get_enm_node(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
node-type of ENM model (either "ca" or "sc" for alpha carbon and side-chain, respectively).
}
\description{
Get ENM node-type
}
