% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_enm_param}
\alias{get_enm_param}
\title{Get ENM parameters}
\usage{
get_enm_param(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
list of ENM parameters
}
\description{
Get ENM parameters
}
