% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfenm_functions.R
\name{get_force}
\alias{get_force}
\title{Get force that mutates wt into mut}
\usage{
get_force(wt, mut)
}
\arguments{
\item{wt}{The wild-type protein}

\item{mut}{The mutant, which contains the mutated \code{graph}}
}
\value{
A force vector of size \code{3 x nsites}
}
\description{
Given \code{mut} obtained by perturbing site edges of \code{wt}, obtain force \code{fij = kij * dlij}
}
\seealso{
Other enm mutating functions: \code{\link{enm_update}},
  \code{\link{get_mutant_site_2}},
  \code{\link{get_mutant_site}}, \code{\link{get_mutant}}
}
\concept{enm mutating functions}
