% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_frustrated}
\alias{get_frustrated}
\title{Get ENM parameter frustrated}
\usage{
get_frustrated(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
frustrated, the parameter that defines whether the ENM is frustrated fully relaxed
}
\description{
Get ENM parameter frustrated
}
