% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_graph}
\alias{get_graph}
\title{Get ENM graph}
\usage{
get_graph(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
graph, a tibble containing the graph representation of the ENM
}
\description{
Get ENM graph
}
