% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_kmat}
\alias{get_kmat}
\title{Get the ENM network matrix K}
\usage{
get_kmat(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
kmat, the 3N x 3N network matrix (N = nsites)
}
\description{
Get the ENM network matrix K
}
