% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_mode}
\alias{get_mode}
\title{Get ENM eigenvector indexes}
\usage{
get_mode(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
a vector of eigenvector indexes
}
\description{
Get ENM eigenvector indexes
}
