% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfenm_functions.R
\name{get_mutant}
\alias{get_mutant}
\title{Get mutant}
\usage{
get_mutant(wt, wt0, ideal = wt0, param)
}
\arguments{
\item{wt}{The protein \code{prot} to mutate}

\item{wt0}{A protein with respect to which mutations are to be calculated}

\item{ideal}{A protein with ideal protein structure}

\item{param}{A list containig \code{enm, mut, fit} parameters}
}
\value{
A list of two components, the mutant and the mutated site: \code{lst(mut, site_mut)}
}
\description{
Gets a mutant by introducing a mutation at a random site
}
\examples{

}
\seealso{
Other enm mutating functions: \code{\link{enm_update}},
  \code{\link{get_force}}, \code{\link{get_mutant_site_2}},
  \code{\link{get_mutant_site}}
}
\concept{enm mutating functions}
