% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm.R
\name{get_mutant_site_lfenm}
\alias{get_mutant_site_lfenm}
\title{Get a single-point mutant using lfenm model}
\usage{
get_mutant_site_lfenm(wt, site_mut, mutation, mut_dl_sigma, mut_sd_min, seed)
}
\arguments{
\item{wt}{The protein \code{prot} to mutate}

\item{site_mut}{The site to mutate (not the pdb_site, but sequential)}

\item{mutation}{An integer, if 0, return \code{wt} without mutating}

\item{mut_dl_sigma}{The standard deviation of a normal distribution from which edge-length perturbation is picked.}

\item{mut_sd_min}{An integer, only edges with \code{sdij > mut_sd_min} are mutated}

\item{seed}{An integer, the seed for set.seed before picking perturbations}
}
\value{
A mutated protein
}
\description{
Returns a mutant given a wt and a site to mutate (site_mut)
}
\examples{

}
\seealso{
Other enm mutating functions: 
\code{\link{get_force}()},
\code{\link{get_mutant_site_sclfenm}()},
\code{\link{get_mutant_site}()}
}
\concept{enm mutating functions}
