% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_pdb_site}
\alias{get_pdb_site}
\title{Get protein sites, pdb numeration}
\usage{
get_pdb_site(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
vector of site indexes (pdb numbers)
}
\description{
Get protein sites, pdb numeration
}
