% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_reduced_kmat}
\alias{get_reduced_kmat}
\title{Calculate reduced ENM K matrix}
\usage{
get_reduced_kmat(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a matrix of size nsites x nsites \eqn{K_{ij} = Tr(\mathbf{K}_{ij})}
}
\description{
Calculates the reduced K matrix (size nsites x nsites)
}
\examples{
wt <- set_enm(pdb_2acy_A, node = "ca", model = "ming_wall",
              d_max = 10.5)

kmat_r <- get_reduced_kmat(wt)
dim(kmat_r)                     # nsites x nsites, from the 3N x 3N kmat

# off-diagonal entries are non-zero only for sites in contact
sum(kmat_r[upper.tri(kmat_r)] != 0)

}
\seealso{
Other site-by-site matrices:
\code{\link[=get_reduced_cmat]{get_reduced_cmat()}},
\code{\link[=get_rho_matrix]{get_rho_matrix()}}
}
\concept{site-by-site matrices}
