% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_site}
\alias{get_site}
\title{Get protein sites}
\usage{
get_site(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
vector of site indexes (numbered from 1 to nsites)
}
\description{
Get protein sites
}
