% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_umat}
\alias{get_umat}
\title{Get ENM eigenvectors}
\usage{
get_umat(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
a matrix of eigenvectors U of size 3 nsites x nmodes
}
\description{
Get ENM eigenvectors
}
