% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_xyz}
\alias{get_xyz}
\title{Get xyz coordinates}
\usage{
get_xyz(prot)
}
\arguments{
\item{prot}{is a prot object}
}
\value{
A 3*N vector of xyz coordinates of the N ENM nodes
}
\description{
Get xyz coordinates
}
