% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prot.R
\name{init_prot}
\alias{init_prot}
\title{Initialize prot object}
\usage{
init_prot(prot, pdb_site_active = NA, ideal = prot, v0 = 0,
  model = "ming_wall", d_max = 10.5, sd_min = 1)
}
\description{
Initialize prot object
}
