% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_add.R
\name{kmat_graph}
\alias{kmat_graph}
\title{Calculate kmat given the ENM graph}
\usage{
kmat_graph(graph, eij, nsites, add_frust = FALSE)
}
\arguments{
\item{graph}{A tibble representing the ENM graph (with edge information, especially \code{kij}}

\item{eij}{A matrix of size \code{n_edges x 3} of \code{eij} versors directed along ENM contacts}

\item{nsites}{The number of nodes of the ENM network}

\item{add_frust=FALSE}{Wether to add frustration or not before calculating \code{kmat}}
}
\value{
The \code{3 nsites x 3 nsites} stiffness matrix of the ENM
}
\description{
Calculate kmat given the ENM graph
}
\examples{
kmat_graph(graph, eij)

}
\seealso{
Other enm builders: \code{\link{enm_add}},
  \code{\link{enm_energy}}, \code{\link{enm_graph_xyz}},
  \code{\link{enm_nma_basis}}, \code{\link{enm_nma}},
  \code{\link{enm_set_xyz}}, \code{\link{enm_set}},
  \code{\link{enm_v0_add}}
}
\concept{enm builders}
