% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{matrix_to_tibble}
\alias{matrix_to_tibble}
\title{turn any matrix into a tibble}
\usage{
matrix_to_tibble(m, row_name = "i", col_name = "j", value_name = "mij")
}
\description{
turn any matrix into a tibble
}
