% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{mode_ensemble_profile}
\alias{mode_ensemble_profile}
\alias{calculate_dmsfn}
\alias{calculate_dhn}
\alias{calculate_rwsipn}
\alias{calculate_nhn}
\title{Compare two protein nsembles mode by mode (mode-dependent profiles)}
\usage{
calculate_dmsfn(wt, mut)

calculate_dhn(wt, mut)

calculate_rwsipn(wt, mut)

calculate_nhn(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{enm} defined}

\item{mut}{A second protein object  with \code{enm} defined}
}
\value{
A vector \code{(x_n)} of size \code{nmodes}, where \code{x_n} is the property compared, for mode n.
}
\description{
Given two proteins, compare their conformational ensembles (fluctuation patterns),
the proteins'  \code{cmat}, and normal modes (Principal components) are assumed known.
}
\details{
(This version works only for wt and mut with no indels)

`calculate_dmsfn` returns mode-dependent profile of changes of mean-square fluctuations \eqn{\delta \sigma_n^2}

`calculate_dhn` returns mode-dependent profile of entropy differences \eqn{\delta H_n}

`calculate_rwsipn` returns mode-dependent profile of rwsip similarity

`calculate_nHn` returns mode-dependent profile of mode conservation  measure \eqn{nH_n}
}
\keyword{internal}
