% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{mode_profile}
\alias{mode_profile}
\alias{calculate_dr2n}
\alias{calculate_de2n}
\alias{calculate_df2n}
\title{Compare two proteins in nm representation}
\usage{
calculate_dr2n(wt, mut)

calculate_de2n(wt, mut)

calculate_df2n(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{xyz} and \code{enm} defined}

\item{mut}{A second protein object  with \code{xyz} defined}
}
\value{
A vector with contributions of each normal mode to the given property
}
\description{
(This version works only for wt and mut with no indels)
}
\details{
`calculate_dr2n` calculates de square of the mode-contributions to \eqn{\delta \mathbf{r} = \mathbf{C}\mathbf{f}}

`calculate_de2n` calculates de square of the mode-contributions to \eqn{\delta \mathbf{e} = \mathbf{C}^{1/2}\mathbf{f}}

`calculate_df2n` calculates de square of the mode-contributions to the force vecgtor \eqn{\mathbf{f}}
}
