% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_mrs.R
\name{mrs_all}
\alias{mrs_all}
\title{Calculate all response matrices and profiles, "simulation" prs method}
\usage{
mrs_all(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min, seed)
}
\description{
Calculate all response matrices and profiles, "simulation" prs method
}
