% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_motion_by_site.R
\name{mrs_motion_by_site}
\alias{mrs_motion_by_site}
\alias{mrs_motion_dmsfij}
\alias{mrs_motion_dhij}
\alias{mrs_motion_rwsipij}
\alias{mrs_motion_dbhatij}
\title{Calculate site-dependent motion-response matrices}
\usage{
mrs_motion_dmsfij(mutants)

mrs_motion_dhij(mutants)

mrs_motion_rwsipij(mutants)

mrs_motion_dbhatij(mutants)
}
\arguments{
\item{mutants}{A tibble of single-point mutants generated using `generate_mutants`}
}
\value{
a response matrix of the form \eqn{R_{ij}} (response site is \code{i}, mutated site is \code{j})
}
\description{
Calculate site-dependent motion-response matrices
}
\details{
`mrs_motion_dmsfij()` calculates the change of msf of site i due to mutations at j, averaged over mutations at \code{j}

`mrs_motion_dhij()` calculates the change in entropy of site i due to mutations at j averaged over mutations at \code{j}

`mrs_motion_rwsipij()` calculates rwsip between mutant and wt site i distributions due to mutations at j, averaged over mutations at j

`mrs_motion_dbhatij()` calculates dbhat distance between wt and mut site i distributions, averaged over mutations at j.
}
