% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_structure_by_mode.R
\name{mrs_structure_by_mode}
\alias{mrs_structure_by_mode}
\alias{mrs_structure_df2nj}
\alias{mrs_structure_de2nj}
\alias{mrs_structure_dr2nj}
\title{Calculate mode-by-site structure-response matrices}
\usage{
mrs_structure_df2nj(mutants)

mrs_structure_de2nj(mutants)

mrs_structure_dr2nj(mutants)
}
\arguments{
\item{mutants}{A tibble of single-point mutants generated using `generate_mutants`}
}
\value{
a response matrix of the form \eqn{R_{nj}} (response mode is \code{n}, mutated site is \code{j})
}
\description{
Calculate mode-by-site structure-response matrices
}
\details{
`mrs_structure_df2nj()` calculates the force matrix \code{f2(n, j)} averaged over mutations at \code{j}

`mrs_structure_de2nj()` calculates the energy matrix \code{de2(n, j)} averaged over mutations at \code{j}

`mrs_structure_dr2nj()` calculates the structural-difference matrix \code{dr2(n, j)} averaged over mutations at \code{j}
}
