% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_structure_by_site.R
\name{mrs_structure_by_site}
\alias{mrs_structure_by_site}
\alias{mrs_structure_df2ij}
\alias{mrs_structure_de2ij}
\alias{mrs_structure_dr2ij}
\alias{mrs_structure_dvsij}
\title{Calculate site-by-site structure-response matrices}
\usage{
mrs_structure_df2ij(mutants)

mrs_structure_de2ij(mutants)

mrs_structure_dr2ij(mutants)

mrs_structure_dvsij(mutants)
}
\arguments{
\item{mutants}{A tibble of single-point mutants generated using `generate_mutants`}
}
\value{
a response matrix of the form \eqn{R_{ij}} (response site is \code{i}, mutated site is \code{j})
}
\description{
Calculate site-by-site structure-response matrices
}
\details{
`mrs_structure_df2ij()` calculates the force matrix \code{df2(i, j)} averaged over mutations at \code{j}

`mrs_structure_de2ij()` calculates the energy-difference matrix \code{de2(i, j)} averaged over mutations at \code{j}

`mrs_structure_dr2ij()` calculates the structural difference matrix \code{dr2(i, j)} averaged over mutations at \code{j}

`mrs_structure_dvsij()` calculates the stress-energy matrix \code{dvs(i, j)} averaged over mutations at \code{j}
}
