% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{msf_site_mode_matrix}
\alias{msf_site_mode_matrix}
\title{MSF of each site contributed by each mode, as matrix}
\usage{
msf_site_mode_matrix(prot)
}
\description{
MSF of each site contributed by each mode, as matrix
}
