% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfenm_functions.R
\name{mutate_eij}
\alias{mutate_eij}
\title{mutate eij vectors, following change of structure and graph}
\usage{
mutate_eij(mut)
}
\description{
mutate eij vectors, following change of structure and graph
}
