% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm.R
\name{mutate_enm}
\alias{mutate_enm}
\title{mutate enm following change in protein structure and lij parameters}
\usage{
mutate_enm(prot, model, d_max, frustrated)
}
\description{
mutate enm following change in protein structure and lij parameters
}
