% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm.R
\name{mutate_graph}
\alias{mutate_graph}
\title{mutate graph following change in structure}
\usage{
mutate_graph(prot)
}
\description{
WARNING: I'm not sure that "frustrated" case is handled well, or in agreement with calc_kmat, etc.
}
\seealso{
Other enm mutating functions: 
\code{\link{get_mutant_site}()}
}
\concept{enm mutating functions}
