% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdb.R
\name{pdb_file_check}
\alias{pdb_file_check}
\title{Check pdb file issues}
\usage{
pdb_file_check(pdb_id, chain = NA_character_, path, prefix = "")
}
\arguments{
\item{pdb_id}{The pdb id (lower case)}

\item{chain}{pdb chain}

\item{path}{path of pdb files}

\item{prefix}{prefix of pdb file names}
}
\value{
If pdb file does not exist, returns NA, otherwise,
    it returns \code{list(missing_residues, n_sites, n_unique_resno)}
}
\description{
Checks if the pdb file exists and, if it does, some other issues
}
\examples{

}
\seealso{
Other pdb input functions: \code{\link{read_pdb_ca}},
  \code{\link{read_pdb_sc}}
}
\concept{pdb input functions}
