% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{prs.new}
\alias{prs.new}
\title{calcualte all fast response matrices and profiles}
\usage{
prs.new(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min, seed = 241956)
}
\description{
calcualte all fast response matrices and profiles
}
