% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{prs.old}
\alias{prs.old}
\title{calcualte all fast response matrices and profiles}
\usage{
prs.old(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min)
}
\description{
calcualte all fast response matrices and profiles
}
