% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{prs}
\alias{prs}
\title{Calculate perturbation-response-scanning responses}
\usage{
prs(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min, beta)
}
\description{
Calculate perturbation-response-scanning responses
}
