% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{prs_all.fast}
\alias{prs_all.fast}
\title{Calculate all response matrices and profiles, "fast prs" method}
\usage{
prs_all.fast(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate all response matrices and profiles, "fast prs" method
}
