% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{prs_all.new}
\alias{prs_all.new}
\title{Calcualte all response matrices and profiles, "new" prs method}
\usage{
prs_all.new(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min, seed)
}
\description{
Calcualte all response matrices and profiles, "new" prs method
}
