% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{prs_all.sim}
\alias{prs_all.sim}
\title{Calculate all response matrices and profiles, "simulation" prs method}
\usage{
prs_all.sim(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min, seed)
}
\description{
Calculate all response matrices and profiles, "simulation" prs method
}
