% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdb.R
\name{read_pdb_ca}
\alias{read_pdb_ca}
\title{Read pdb alpha carbons}
\usage{
read_pdb_ca(pdb_id, chain = NA_character_, path, prefix = "")
}
\arguments{
\item{pdb_id}{protein's pdb id}

\item{chain}{protein chain}

\item{path}{path to pdb files}

\item{prefix}{prefix of pdb file names}
}
\value{
list of center-of-mass coordinates lst(xyz, pdb_site, bfactor)
}
\description{
Reads pdb file, returns alpha-carbon coordinates and bfactors
}
\examples{

}
\seealso{
Other pdb input functions: \code{\link{pdb_file_check}},
  \code{\link{read_pdb_sc}}
}
\concept{pdb input functions}
