% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmrs.R
\name{sdmrs}
\alias{sdmrs}
\title{Calculate double-mutational-scanning matrix}
\usage{
sdmrs(wt, nmut, mut_dl_sigma, mut_sd_min, seed, option = "mean_max")
}
\description{
Calculate double-mutational-scanning matrix
}
