% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prot.R
\name{set_prot}
\alias{set_prot}
\title{Set up ENM from bio3d pdb object}
\usage{
set_prot(pdb, node = "sc", model, d_max, frustrated)
}
\arguments{
\item{pdb}{a pdb object obtained using bio3d::read.pdb}

\item{node}{network nodes: "sc" or "ca"}

\item{model}{enm model: "anm", "ming_wall", "hnm", "hnm0", "pfanm", "reach"}

\item{d_max}{cutoff to define enm contacts}

\item{frustrated}{logical indicating whether to include frustrations in calculation of kmat}
}
\value{
a `prot` object (list containing xyz, site, etc.)
}
\description{
Set up ENM from bio3d pdb object
}
\examples{

}
