% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{site_ensemble_profile}
\alias{site_ensemble_profile}
\alias{calculate_dmsfi}
\alias{calculate_dbhati}
\alias{calculate_rwsipi}
\alias{calculate_dhi}
\title{Compare two protein nsembles site by site (site-dependent profiles)}
\usage{
calculate_dmsfi(wt, mut)

calculate_dbhati(wt, mut)

calculate_rwsipi(wt, mut)

calculate_dhi(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{enm} defined}

\item{mut}{A second protein object  with \code{enm} defined}
}
\value{
A vector \code{(x_i)} of size \code{nsites}, where \code{x_i} is the property compared, for site i.
}
\description{
Given two proteins, compare their conformational ensembles (fluctuation patterns),
the proteins'  \code{cmat}, and normal modes (Principal components) are assumed known.
}
\details{
(This version works only for wt and mut with no indels)

`calculate_dmsfi` returns site-dependent profile of changes of mean-square fluctuations \eqn{\delta \sigma_i^2}

`calculate_dbhati` returns site-dependent profile of dbhat distances

`calculate_rwsipi` returns site-dependent profile of rwsip similarity

`calculate_dhi` returns site-dependent profile of dh similarity
}
