% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{site_profile}
\alias{site_profile}
\alias{calculate_dr2i}
\alias{calculate_de2i}
\alias{calculate_df2i}
\alias{calculate_dvmi}
\alias{calculate_dvsi}
\alias{calculate_dvsi_same_topology}
\title{Compare two proteins site by site (site-dependent profiles)}
\usage{
calculate_dr2i(wt, mut)

calculate_de2i(wt, mut, kmat_sqrt)

calculate_df2i(wt, mut)

calculate_dvmi(wt, mut)

calculate_dvsi(wt, mut)

calculate_dvsi_same_topology(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{xyz} defined}

\item{mut}{A second protein object  with \code{xyz} defined}
}
\value{
A vector \code{(x_i)} of size \code{nsites}, where \code{x_i} is the property compared, for site i.
}
\description{
This version works only for wt and mut with no indels
}
\details{
`calculate_dr2i` returns the square of structural difference vector \eqn{\mathbf{C}\mathbf{f}}

`calculate_de2i` returns the square of deformation energy vector \eqn{\mathbf{C}^{1/2}\mathbf{f}}

`calculate_df2i` returns the square of force vector \eqn{\mathbf{f}}

`calculate_dvmi` returns the difference of site-dependent minimum-energy contributions

`calculate_dvsi` returns the difference of site-dependent stress-energy contributions

`calculate_dvsi_same_topology` returns the difference of site-dependent stress-energy contributions, assumes no change in topology
}
