% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_smrs.R
\name{smrs_all}
\alias{smrs_all}
\title{Calcualte all response matrices and profiles, "new" prs method}
\usage{
smrs_all(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min, seed)
}
\description{
Calcualte all response matrices and profiles, "new" prs method
}
