% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_exported.R
\name{beta_boltzmann}
\alias{beta_boltzmann}
\title{Calculate's Bolzmann's beta = 1 / RT}
\usage{
beta_boltzmann(R = 0.001986, T = 298)
}
\arguments{
\item{R}{is Boltzmann's constant per mol}

\item{T}{is absolute temperature in Kelving}
}
\value{
1 / (RT)
}
\description{
Calculate's Bolzmann's beta = 1 / RT
}
