% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_sim.R
\name{calculate_df2ij.prs}
\alias{calculate_df2ij.prs}
\title{Calculate site-by-site force matrix df2(i, j) using method "sim"}
\usage{
calculate_df2ij.prs(mutants)
}
\description{
Calculate site-by-site force matrix df2(i, j) using method "sim"
}
