% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_mrs.R
\name{calculate_df2nj.prs}
\alias{calculate_df2nj.prs}
\title{Calculate mode-by-site force response matrix df2(n, j) using method "sim"}
\usage{
calculate_df2nj.prs(mutants)
}
\description{
Calculate mode-by-site force response matrix df2(n, j) using method "sim"
}
