% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_fast.R
\name{calculate_dfij.fast}
\alias{calculate_dfij.fast}
\title{Calculate mutational response matrices, site-by-site}
\usage{
calculate_dfij.fast(wt, mut_dl_sigma, mut_sd_min)
}
\description{
Calculate mutational response matrices, site-by-site
}
