% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{calculate_dr2i}
\alias{calculate_dr2i}
\alias{calculate_de2i}
\alias{calculate_df2i}
\title{Compare the structures of two proteins in site representation}
\usage{
calculate_dr2i(wt, mut)

calculate_de2i(wt, mut, kmat_sqrt)

calculate_df2i(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{xyz} defined}

\item{mut}{A second protein object  with \code{xyz} defined}
}
\value{
A tibble with columns \code{pdb_site, site, dr2}
}
\description{
Calculate de difference between the structures of two proteins, return dr2i
}
\details{
This version works only for wt and mut with no indels
}
