% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{calculate_dr2n}
\alias{calculate_dr2n}
\alias{calculate_de2n}
\alias{calculate_df2n}
\title{Compare the structures of two proteins in nm representation}
\usage{
calculate_dr2n(wt, mut)

calculate_de2n(wt, mut)

calculate_df2n(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{xyz} and \code{enm} defined}

\item{mut}{A second protein object  with \code{xyz} defined}
}
\value{
A vector of square differences along normal modes \code{dr2n}
}
\description{
Calculate de difference between the structures of two proteins, return dr2n.
}
\details{
This version works only for wt and mut with no indels
}
