% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{calculate_dvs}
\alias{calculate_dvs}
\title{Stress-model difference of local-mutational-stress energy}
\usage{
calculate_dvs(prot1, prot2, ideal = prot1)
}
\description{
Stress-model difference of local-mutational-stress energy
}
