% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_sim.R
\name{calculate_dvsij.prs}
\alias{calculate_dvsij.prs}
\title{Calculate site-by-site stress energy response matrix dvs(i, j) using method "sim"}
\usage{
calculate_dvsij.prs(mutants)
}
\description{
Calculate site-by-site stress energy response matrix dvs(i, j) using method "sim"
}
