% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_new.R
\name{calculate_dvsjm.new}
\alias{calculate_dvsjm.new}
\title{calcualte stress energy change vector dvsjm}
\usage{
calculate_dvsjm.new(wt, nmut_per_site, mut_dl_sigma, mut_sd_min, seed = 241956)
}
\description{
dvsjm is dvs at site j, for mutations m = c(1, 2, 3, ..., nmut_per_site)
}
