% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{calculate_enm_kmat}
\alias{calculate_enm_kmat}
\title{Calculate kmat given the ENM graph}
\usage{
calculate_enm_kmat(graph, eij, nsites, frustrated)
}
\arguments{
\item{graph}{A tibble representing the ENM graph (with edge information, especially \code{kij}}

\item{eij}{A matrix of size \code{n_edges x 3} of \code{eij} versors directed along ENM contacts}

\item{nsites}{The number of nodes of the ENM network}

\item{frustrated}{Logical indicating  wether to add frustration or not before calculating \code{kmat}}
}
\value{
The \code{3 nsites x 3 nsites} stiffness matrix of the ENM
}
\description{
Calculate kmat given the ENM graph
}
\examples{
calculate_enm_kmat(graph, eij)

}
\seealso{
Other enm builders: 
\code{\link{calculate_enm_graph}()},
\code{\link{calculate_enm_nma}()}
}
\concept{enm builders}
