% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{calculate_enm_nma}
\alias{calculate_enm_nma}
\title{Perform Normal Mode Analysis}
\usage{
calculate_enm_nma(kmat, too_small = 1e-05)
}
\arguments{
\item{kmat}{The K matrix to diagonalize}

\item{too_small=1.e-5}{A small value, eigenvectors with eigenvalues larger than `too_small` are discarded}
}
\value{
A list with elements \code{lst(mode,evalue,cmat,umat)}
}
\description{
Given an enm `kmat`, perform NMA
}
\examples{
\dontrun{
calculate_enm_anm(kmat, too_small = 1.e-10)
}

}
\seealso{
Other enm builders: 
\code{\link{calculate_enm_eij}()},
\code{\link{calculate_enm_graph}()},
\code{\link{calculate_enm_kmat}()},
\code{\link{calculate_enm_nodes}()}
}
\concept{enm builders}
\keyword{internal}
