% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm_analysis.R
\name{delta_energy}
\alias{delta_energy}
\alias{calculate_dvm}
\alias{calculate_dg_entropy}
\alias{calculate_dvs}
\title{Calculate energy differences between a mutant and wild type}
\usage{
calculate_dvm(wt, mut)

calculate_dg_entropy(wt, mut, beta)

calculate_dvs(wt, mut, ideal = wt)
}
\arguments{
\item{wt}{A protein object with \code{xyz} defined}

\item{mut}{A second protein object  with \code{xyz} defined}
}
\value{
A (scalar) energy difference between mutant and wild type.
}
\description{
Calculate energy differences between a mutant and wild type
}
\details{
`calculate_dvm` calculates the minimum-energy difference between \code{mut} and \code{wt}

`calculate_dg_enetropy` calculates the entropic free energ difference between \code{mut} and \code{wt}

`calculate_dvs` calculates the ideal-conformation stress-energy difference between \code{mut} and \code{wt}
}
