% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_motion_by_mode.R
\name{delta_motion_by_mode}
\alias{delta_motion_by_mode}
\alias{delta_motion_dmsfn}
\alias{delta_motion_dhn}
\alias{delta_motion_rwsipn}
\alias{delta_motion_nhn}
\title{Compare two protein ensembles mode by mode (mode-dependent profiles)}
\usage{
delta_motion_dmsfn(wt, mut)

delta_motion_dhn(wt, mut)

delta_motion_rwsipn(wt, mut)

delta_motion_nhn(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{enm} defined}

\item{mut}{A second protein object  with \code{enm} defined}
}
\value{
A vector \code{(x_n)} of size \code{nmodes}, where \code{x_n} is the property compared, for mode n.
}
\description{
Given two proteins, compare their conformational ensembles (fluctuation patterns),
the proteins'  \code{cmat}, and normal modes (Principal components) are assumed known.
}
\details{
(This version works only for wt and mut with no indels)

`delta_motion_dmsfn` returns mode-dependent profile of changes of mean-square fluctuations \eqn{\delta \sigma_n^2}

`delta_motion_dhn` returns mode-dependent profile of entropy differences \eqn{\delta H_n}

`delta_motion_rwsipn` returns mode-dependent profile of rwsip similarity

`delta_motion_nhn` returns mode-dependent profile of mode conservation  measure \eqn{nH_n}
}
