% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_motion_by_site.R
\name{delta_motion_by_site}
\alias{delta_motion_by_site}
\alias{delta_motion_dmsfi}
\alias{delta_motion_dbhati}
\alias{delta_motion_rwsipi}
\alias{delta_motion_dhi}
\title{Compare the motions of two proteins site by site}
\usage{
delta_motion_dmsfi(wt, mut)

delta_motion_dbhati(wt, mut)

delta_motion_rwsipi(wt, mut)

delta_motion_dhi(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{enm} defined}

\item{mut}{A second protein object  with \code{enm} defined}
}
\value{
A vector \code{(x_i)} of size \code{nsites}, where \code{x_i} is the property compared, for site i.
}
\description{
Given two proteins, compare their conformational ensembles (fluctuation patterns),
the proteins'  \code{cmat}, and normal modes (Principal components) are assumed known.
}
\details{
(This version works only for wt and mut with no indels)

`delta_motion_dmsfi` returns site-dependent profile of changes of mean-square fluctuations \eqn{\delta \sigma_i^2}

`delta_motion_dbhati` returns site-dependent profile of dbhat distances

`delta_motion_rwsipi` returns site-dependent profile of rwsip similarity

`delta_motion_dhi` returns site-dependent profile of dh similarity
}
