% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_structure_by_mode.R
\name{delta_structure_by_mode}
\alias{delta_structure_by_mode}
\alias{delta_structure_dr2n}
\alias{delta_structure_de2n}
\alias{delta_structure_df2n}
\title{Compare two protein structures in nm representation}
\usage{
delta_structure_dr2n(wt, mut)

delta_structure_de2n(wt, mut)

delta_structure_df2n(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{xyz} and \code{enm} defined}

\item{mut}{A second protein object  with \code{xyz} defined}
}
\value{
A vector with contributions of each normal mode to the given property
}
\description{
(This version works only for wt and mut with no indels)
}
\details{
`delta_structure_dr2n` calculates de square of the mode-contributions to \eqn{\delta \mathbf{r} = \mathbf{C}\mathbf{f}}

`delta_structure_de2n` calculates de square of the mode-contributions to \eqn{\delta \mathbf{e} = \mathbf{C}^{1/2}\mathbf{f}}

`delta_structure_df2n` calculates de square of the mode-contributions to the force vecgtor \eqn{\mathbf{f}}
}
