% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy_activation.R
\name{dgact_dv}
\alias{dgact_dv}
\title{Activation free energy, internal energy contribution}
\usage{
dgact_dv(prot, ideal, pdb_site_active = NA)
}
\description{
Activation free energy, internal energy contribution
}
\seealso{
Other enm_energy: 
\code{\link{dgact_tds}()},
\code{\link{enm_g_entropy}()},
\code{\link{enm_v_min}()}
}
\concept{enm_energy}
