% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy_activation.R
\name{dgact_tds}
\alias{dgact_tds}
\title{Activation free energy, entropic contribution}
\usage{
dgact_tds(prot, ideal, pdb_site_active = NA, beta = beta_boltzmann())
}
\arguments{
\item{prot}{A protein object obtained using \code{set_enm()}}

\item{ideal}{A protein object whose conformation defines the ideal (transition) state}

\item{pdb_site_active}{A vector of active-site residues, in pdb numbering (\code{resno}).
If \code{NA} (the default), the function returns \code{NA}.}

\item{beta}{Inverse temperature, \code{1/kT}}
}
\value{
A scalar, the entropic contribution to the activation free energy,
  or \code{NA} if \code{pdb_site_active} is \code{NA}.
}
\description{
Calculates the entropic contribution to the activation free energy, as the
difference between the entropic free energy of the active site of \code{prot} and
that of the transition state, which is assumed rigid (contributing zero).
}
\seealso{
Other enm_energy:
\code{\link[=beta_boltzmann]{beta_boltzmann()}},
\code{\link[=dgact_dv]{dgact_dv()}},
\code{\link[=enm_g_entropy]{enm_g_entropy()}},
\code{\link[=enm_v_min]{enm_v_min()}}
}
\concept{enm_energy}
