% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy_activation.R
\name{dgact_tds}
\alias{dgact_tds}
\title{Activation free energy, entropic contribution}
\usage{
dgact_tds(prot, ideal, pdb_site_active = NA, beta = beta_boltzmann())
}
\description{
Activation free energy, entropic contribution
}
\seealso{
Other enm_energy: 
\code{\link{dgact_dv}()},
\code{\link{enm_g_entropy}()},
\code{\link{enm_v_min}()}
}
\concept{enm_energy}
