% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy.R
\name{enm_g_entropy}
\alias{enm_g_entropy}
\title{Calculate entropic total free energy of prot object}
\usage{
enm_g_entropy(prot, beta)
}
\arguments{
\item{prot}{is a prot object with known eigenvalues of enm model}

\item{beta}{is 1 / kT}
}
\description{
Calculate entropic total free energy of prot object
}
\seealso{
Other enm_energy:
\code{\link[=dgact_dv]{dgact_dv()}},
\code{\link[=dgact_tds]{dgact_tds()}},
\code{\link[=enm_v_min]{enm_v_min()}}
}
\concept{enm_energy}
