% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_nma.R
\name{enm_nma_basis}
\alias{enm_nma_basis}
\title{NMA of ENM using a basis}
\usage{
enm_nma_basis(kmat, umat0, nbasis = ncol(umat0), TOL = 1e-05)
}
\arguments{
\item{kmat}{An enm K matrix}

\item{umat0}{A matrix of basis vectors (e.g. normal modes of unperturbed enm)}

\item{nbasis}{Number of basis vectors to use (must be smaller than `ncol(umat0)`)}

\item{TOL}{A number near 0 to use to discard normal modes (eval > TOL are kept)}
}
\value{
A list containing `mode, evalue, cmat, umat`
}
\description{
Before diagonalizing `kmat`, transform into a (smaller) basis. Useful for normal-mode perturbation calculations
}
\examples{

}
\seealso{
Other enm builders: 
\code{\link{add_enm_activation}()},
\code{\link{enm_nma}()},
\code{\link{set_enm_graph}()},
\code{\link{set_enm_kmat}()}
}
\concept{enm builders}
