% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_add.R
\name{enm_set}
\alias{enm_set}
\title{Set initial \code{enm} object}
\usage{
enm_set(prot, param, ...)
}
\arguments{
\item{prot}{A protein object that must contain \code{xyz} and \code{pdb_site} elements}

\item{param}{A list of \code{enm} parameters including elements \code{model, k_scale, d_max, sd_min}}

\item{...}{Other parameters}
}
\value{
A list containing the ENM graph `graph`, the matrix of versors `eij`, and the ENM K matrix `kmat`
}
\description{
Set initial \code{enm} object
}
\seealso{
Other enm builders: \code{\link{enm_add}},
  \code{\link{enm_energy}}, \code{\link{enm_graph_xyz}},
  \code{\link{enm_nma_basis}}, \code{\link{enm_nma}},
  \code{\link{enm_set_xyz}}, \code{\link{enm_v0_add}},
  \code{\link{kmat_graph}}
}
\concept{enm builders}
