% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy.R
\name{enm_v_min}
\alias{enm_v_min}
\title{Calculate minimum energy of a given prot object}
\usage{
enm_v_min(prot)
}
\arguments{
\item{prot}{is a prot object, wit a component graph tibble
where v0ij, kij, lij and the dij for the minimum conformation are found.}
}
\value{
a scalar: the energy at the minimum-energy conformation
}
\description{
Calculate minimum energy of a given prot object
}
\seealso{
Other enm_energy:
\code{\link[=dgact_dv]{dgact_dv()}},
\code{\link[=dgact_tds]{dgact_tds()}},
\code{\link[=enm_g_entropy]{enm_g_entropy()}}
}
\concept{enm_energy}
