% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy.R
\name{enm_v_min}
\alias{enm_v_min}
\title{Calculate minimum energy of a given prot object}
\usage{
enm_v_min(prot)
}
\arguments{
\item{prot}{is a prot object with a en enm$graph tibble
where v0ij, kij, lij and the dij for the minimum conformation are found.}
}
\value{
a scalar: the minimum energy
}
\description{
Calculate minimum energy of a given prot object
}
