% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{generate_mutants}
\alias{generate_mutants}
\title{Generarte table of single-point mutants}
\usage{
generate_mutants(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min, seed)
}
\description{
Generarte table of single-point mutants
}
