% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_cmat_sqrt}
\alias{get_cmat_sqrt}
\title{Calculate the matrix square root of the ENM covariance matrix}
\usage{
get_cmat_sqrt(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a matrix of size 3 nsites x 3 nsites, the matrix square root of \code{cmat}
}
\description{
Calculates \eqn{\mathbf{C}^{1/2}} from the eigendecomposition of the network,
as \eqn{\mathbf{U} \sqrt{1/\lambda} \mathbf{U}^T}. The dual of
\code{\link{get_kmat_sqrt}}, which uses \eqn{\sqrt{\lambda}} instead.
}
\seealso{
Other matrix square roots:
\code{\link[=get_kmat_sqrt]{get_kmat_sqrt()}}
}
\concept{matrix square roots}
