% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_cn}
\alias{get_cn}
\title{Calculate CN site-dependent profile}
\usage{
get_cn(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a vector of size nsites with cn values for each site
}
\description{
Calculates the Contact Number (CN) of each site
}
\examples{
wt <- set_enm(pdb_2acy_A, node = "ca", model = "ming_wall",
              d_max = 10.5, frustrated = FALSE)

cn <- get_cn(wt)
length(cn)                      # one value per site
which.max(cn)                   # most buried site by contact number

}
\seealso{
Other site profiles:
\code{\link[=get_dactive]{get_dactive()}},
\code{\link[=get_mlms]{get_mlms()}},
\code{\link[=get_msf_site]{get_msf_site()}},
\code{\link[=get_stress]{get_stress()}},
\code{\link[=get_wcn]{get_wcn()}}
}
\concept{site profiles}
