% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_cn}
\alias{get_cn}
\title{Calculate CN site-dependent profile}
\usage{
get_cn(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a vector of size nsites with cn values for each site
}
\description{
Calculates the Contact Number (CN) of each site
}
\seealso{
Other prot getters: 
\code{\link{get_mlms}()},
\code{\link{get_msf_mode}()},
\code{\link{get_msf_site_mode}()},
\code{\link{get_msf_site}()},
\code{\link{get_reduced_cmat}()},
\code{\link{get_reduced_kmat}()},
\code{\link{get_rho_matrix}()},
\code{\link{get_stress}()},
\code{\link{get_umat2}()}
}
\concept{prot getters}
