% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_dactive}
\alias{get_dactive}
\title{Calculate distance to active site}
\usage{
get_dactive(prot, pdb_site_active)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}

\item{pdb_site_active}{is a vector of pdb resno of active residues.}
}
\value{
a vector of size nsites with dactive values for each site
}
\description{
Calculates the distance from each site to the closest active residue
}
\seealso{
Other site profiles: 
\code{\link{get_cn}()},
\code{\link{get_mlms}()},
\code{\link{get_msf_site}()},
\code{\link{get_stress}()},
\code{\link{get_wcn}()}
}
\concept{site profiles}
