% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm.R
\name{get_force}
\alias{get_force}
\title{Get force resulting from adding delta_lij to wt}
\usage{
get_force(wt, delta_lij)
}
\arguments{
\item{wt}{the wild-type protein}

\item{delta_lij}{the perturbations to the wt lij parameters}
}
\value{
A force vector of size \code{3 x nsites}
}
\description{
Get force resulting from adding delta_lij to wt
}
\seealso{
Other enm mutating functions: 
\code{\link{get_mutant_site}()}
}
\concept{enm mutating functions}
