% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_kmat_sqrt}
\alias{get_kmat_sqrt}
\title{Calculate the matrix square root of the ENM K matrix}
\usage{
get_kmat_sqrt(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a matrix of size 3 nsites x 3 nsites, the matrix square root of \code{kmat}
}
\description{
Calculates \eqn{\mathbf{K}^{1/2}} from the eigendecomposition of the network,
as \eqn{\mathbf{U} \sqrt{\lambda} \mathbf{U}^T}. A general matrix-square-root
routine is not usable here: \code{kmat} is singular (the six rigid-body modes
have zero eigenvalue), so building it from the modes is what makes it defined.
}
\details{
Needed to call \code{\link{delta_structure_de2i}}, which takes
\code{kmat_sqrt} as an argument.
}
\examples{
wt <- set_enm(pdb_2acy_A, node = "ca", model = "ming_wall",
              d_max = 10.5)

ks <- get_kmat_sqrt(wt)
dim(ks)                         # 3 nsites x 3 nsites

# squaring it returns the network matrix
all.equal(ks \%*\% ks, as.matrix(get_kmat(wt)), check.attributes = FALSE)

}
\seealso{
Other matrix square roots:
\code{\link[=get_cmat_sqrt]{get_cmat_sqrt()}}
}
\concept{matrix square roots}
