% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_mlms}
\alias{get_mlms}
\title{Calculate MLMS site-dependent profile}
\usage{
get_mlms(prot, sdij_cut = 2)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}

\item{sdij_cut}{An integer cutoff of sequence distance to include in calculation}
}
\value{
the profile of mean-local-mutational-stress (mlms) values
}
\description{
Calculates the Mean Local Mutational Stress (MLMS) profile using graph of prot object
}
\examples{
wt <- set_enm(pdb_2acy_A, node = "ca", model = "ming_wall",
              d_max = 10.5)

mlms <- get_mlms(wt)
length(mlms)                    # one value per site

# sdij_cut drops near-in-sequence contacts; raising it keeps fewer springs
sum(get_mlms(wt, sdij_cut = 5)) < sum(mlms)

}
\seealso{
Other site profiles:
\code{\link[=get_cn]{get_cn()}},
\code{\link[=get_dactive]{get_dactive()}},
\code{\link[=get_msf_site]{get_msf_site()}},
\code{\link[=get_wcn]{get_wcn()}}
}
\concept{site profiles}
