% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_msf_mode}
\alias{get_msf_mode}
\title{Calculate MSF mode-dependent profile}
\usage{
get_msf_mode(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a vector of size nsites with msf values for each mode
}
\description{
Calculates the mean-square-fluctuation in the direction of each normal mode
}
\concept{mode profiles}
