% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_msf_site}
\alias{get_msf_site}
\title{Calculate MSF site-dependent profile}
\usage{
get_msf_site(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a vector of size nsites with msf values for each site
}
\description{
Calculates the mean-square-fluctuation of each site
}
\seealso{
Other site profiles:
\code{\link[=get_cn]{get_cn()}},
\code{\link[=get_dactive]{get_dactive()}},
\code{\link[=get_mlms]{get_mlms()}},
\code{\link[=get_stress]{get_stress()}},
\code{\link[=get_wcn]{get_wcn()}}
}
\concept{site profiles}
