% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_msf_site_mode}
\alias{get_msf_site_mode}
\title{MSF of each site contributed by each mode, as matrix}
\usage{
get_msf_site_mode(prot)
}
\description{
MSF of each site contributed by each mode, as matrix
}
