% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfenm_functions.R
\name{get_mutant_site}
\alias{get_mutant_site}
\title{Get a single-point mutant}
\usage{
get_mutant_site(wt, site_mut, wt0, ideal = wt0, param)
}
\arguments{
\item{wt}{The protein \code{prot} to mutate}

\item{site_mut}{The site to mutate (not the pdb_site, but sequential)}

\item{wt0}{A protein with respect to which mutations are to be calculated}

\item{ideal}{A protein with ideal protein structure}

\item{param}{A list containig \code{enm, mut, fit} parameters}
}
\description{
Returns a mutant given a wt and a site to mutate (site_mut)
    wt0 is the protein to use to calculate perturbations of current site
    (if wt0 = wt, it is the current wild-type, but it can be a fixed initial state)
    ideal is the ideal protein (assumed to be by default the fixed initial state)
    copy wt
}
\examples{

}
\seealso{
Other enm mutating functions: \code{\link{enm_update}},
  \code{\link{get_force}}, \code{\link{get_mutant_site_2}},
  \code{\link{get_mutant}}
}
\concept{enm mutating functions}
