% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm.R
\name{get_mutant_site}
\alias{get_mutant_site}
\title{Get a single-point mutant}
\usage{
get_mutant_site(
  wt,
  site_mut,
  mutation = 0,
  mut_model = "lfenm",
  mut_dl_sigma = 0.3,
  mut_sd_min = 2,
  seed = 241956
)
}
\arguments{
\item{wt}{The protein \code{prot} to mutate}

\item{site_mut}{The site to mutate (not the pdb_site, but sequential)}

\item{mutation}{An integer, if 0, return \code{wt} without mutating}

\item{mut_model}{A string specifying mutational model ("lfenm" or "sclfenm")}

\item{mut_dl_sigma}{The standard deviation of a normal distribution from which edge-length perturbation is picked.}

\item{mut_sd_min}{An integer, only edges with \code{sdij >= mut_sd_min} are mutated}

\item{seed}{An integer, the seed for set.seed before picking perturbations}
}
\value{
A mutated protein
}
\description{
Returns a mutant given a wt and a site to mutate (site_mut)
}
\examples{

}
\seealso{
Other enm mutating functions: 
\code{\link{get_force}()},
\code{\link{get_mutant_site_lfenm}()},
\code{\link{get_mutant_site_sclfenm}()}
}
\concept{enm mutating functions}
