% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfenm_functions.R
\name{get_mutant_site}
\alias{get_mutant_site}
\title{Get a single-point mutant}
\usage{
get_mutant_site(wt, site_mut, mutation = 0, seed = 241956 + site_mut *
  mutation, wt0 = wt, ideal = wt0, sd_min = 2, dl_sigma = 0.3,
  model = "ming_wall", d_max = 10.5, add_frust = FALSE,
  update_enm = FALSE)
}
\arguments{
\item{wt}{The protein \code{prot} to mutate}

\item{site_mut}{The site to mutate (not the pdb_site, but sequential)}

\item{mutation}{An integer, if 0, return \code{wt} without mutating}

\item{seed}{An integer, the seed for set.seed before picking perturbations}

\item{wt0}{A protein with respect to which mutations are to be calculated}

\item{ideal}{A protein with ideal protein structure}

\item{sd_min}{An integer, only edges with \code{sdij > sd_min} are mutated}

\item{dl_sigma}{The standard deviation of a normal distribution from which edge-length perturbation is picked.}

\item{model}{The enm model type}

\item{d_max}{The enm distance cut-off to define contacts}

\item{add_frust}{Whether to include frustrations or not}

\item{update_enm}{Whether to update enm or not}
}
\description{
Returns a mutant given a wt and a site to mutate (site_mut)
    wt0 is the protein to use to calculate perturbations of current site
    (if wt0 = wt, it is the current wild-type, but it can be a fixed initial state)
    ideal is the ideal protein (assumed to be by default the fixed initial state)
    copy wt
}
\examples{

}
\seealso{
Other enm mutating functions: \code{\link{enm_update}},
  \code{\link{get_force}}
}
\concept{enm mutating functions}
