% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_getters.R
\name{get_prot_property}
\alias{get_prot_property}
\alias{get_enm_param}
\alias{get_nsites}
\alias{get_site}
\alias{get_pdb_site}
\alias{get_bfactor}
\alias{get_xyz}
\title{Get various properties of a prot object}
\usage{
get_enm_param(prot)

get_nsites(prot)

get_site(prot)

get_pdb_site(prot)

get_bfactor(prot)

get_xyz(prot)
}
\arguments{
\item{prot}{A protein with its associated ENM model, obtained using `set_enm`}
}
\value{
\code{get_enm_param}: list of ENM parameters

\code{get_nsites}: number of sites (ENM nodes)

\code{get_site}: site indexes, from `1` to `nsites`

\code{get_pdb_site}: site indexes, pdb numeration (resno)

\code{get_bfactor}: B-factors of the X-ray file for the ENM nodes

\code{get_xyz}: A 3*N vector of xyz coordinates of the N ENM nodes
}
\description{
Get various properties of a prot object
}
