% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_reduced_cmat}
\alias{get_reduced_cmat}
\title{Calculate reduced covariance matrix}
\usage{
get_reduced_cmat(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a matrix of size nsites x nsites with \eqn{c_{ij} = < d\mathbf{r}_i . d\mathbf{r}_j >}
}
\description{
Calculates the reduced covariance matrix (size nsites x nsites)
}
\seealso{
Other site-by-site matrices:
\code{\link[=get_reduced_kmat]{get_reduced_kmat()}},
\code{\link[=get_rho_matrix]{get_rho_matrix()}}
}
\concept{site-by-site matrices}
