% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_reduced_kmat}
\alias{get_reduced_kmat}
\title{Calculate reduced ENM K matrix}
\usage{
get_reduced_kmat(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a matrix of size nsites x nsites \eqn{K_{ij} = Tr(\mathbf{K}_{ij})}
}
\description{
Calculates the reduced K matrix (size nsites x nsites)
}
\seealso{
Other site-by-site matrices:
\code{\link[=get_reduced_cmat]{get_reduced_cmat()}},
\code{\link[=get_rho_matrix]{get_rho_matrix()}}
}
\concept{site-by-site matrices}
