% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_umat2}
\alias{get_umat2}
\title{Calculate Reduced \code{umat^2}}
\usage{
get_umat2(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a matrix of size nsites x nmodes with contribution of each site to each mode.
}
\description{
Calculates a matrix of size nsites x nmodes. Element umat2(i,n) is the contribution of site i to mode n (amplitude squared, added over x,y,z)
}
\seealso{
Other site-by-mode matrices:
\code{\link[=get_msf_site_mode]{get_msf_site_mode()}}
}
\concept{site-by-mode matrices}
