% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_wcn}
\alias{get_wcn}
\title{Calculate WCN site-dependent profile}
\usage{
get_wcn(prot)
}
\arguments{
\item{prot}{is a protein object obtained using set_enm()}
}
\value{
a vector of size nsites with wcn values for each site
}
\description{
Calculates the Weighted Contact Number (WCN) of each site
}
\seealso{
Other site profiles:
\code{\link[=get_cn]{get_cn()}},
\code{\link[=get_dactive]{get_dactive()}},
\code{\link[=get_mlms]{get_mlms()}},
\code{\link[=get_msf_site]{get_msf_site()}},
\code{\link[=get_stress]{get_stress()}}
}
\concept{site profiles}
