% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm-package.R
\docType{package}
\name{penm-package}
\alias{penm}
\alias{penm-package}
\title{penm: Build and Perturb Elastic Network Models}
\description{
Functions to calculate ENM models, mutate ENMs by perturbing them, and measure the resulting differences between wild-type and mutant in energy, structure, and motion.
}
\details{
The \code{penm} package includes functions to calculate various Elastic Network Models for proteins,
    perform normal mode analysis, and using lfenm, obtain mutant proteins and the corresponding mutant ENMs.
    In addition, it has functions to scan the various average-responses w.r.t. single-site mutations and double-site mutations.
}
\author{
\strong{Maintainer}: Julian Echave \email{jechave@unsam.edu.ar} (\href{https://orcid.org/0000-0003-0678-5593}{ORCID})

Authors:
\itemize{
  \item Julian Echave \email{jechave@unsam.edu.ar} (\href{https://orcid.org/0000-0003-0678-5593}{ORCID})
}

}
