% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm-package.R
\docType{package}
\name{penm-package}
\alias{penm}
\alias{penm-package}
\title{penm: Build and Perturb Elastic Network Models}
\description{
Functions to calculate ENMs of proteins and their mutants
}
\details{
For proteins, the \code{penm} package includes functions to calculate various Elastic Network Models,
    perform normal mode analysis, and using lfenm, obtain mutant proteins and the corresponding mutant ENMs.
}
\keyword{internal}
