% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{prs}
\alias{prs}
\title{calcualte all fast response matrices and profiles}
\usage{
prs(wt, nmut_per_site, mut_model, mut_dl_sigma, mut_sd_min)
}
\description{
calcualte all fast response matrices and profiles
}
