% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_mrs.R
\name{site_mrs_matrices}
\alias{site_mrs_matrices}
\alias{calculate_df2ij_mrs}
\alias{calculate_de2ij_mrs}
\alias{calculate_dr2ij_mrs}
\alias{calculate_dvsij_mrs}
\alias{calculate_dmsfij_mrs}
\alias{calculate_dhij_mrs}
\alias{calculate_rwsipij_mrs}
\alias{calculate_dbhatij_mrs}
\title{Calculate site-dependent structural response matrices}
\usage{
calculate_df2ij_mrs(mutants)

calculate_de2ij_mrs(mutants)

calculate_dr2ij_mrs(mutants)

calculate_dvsij_mrs(mutants)

calculate_dmsfij_mrs(mutants)

calculate_dhij_mrs(mutants)

calculate_rwsipij_mrs(mutants)

calculate_dbhatij_mrs(mutants)
}
\arguments{
\item{mutants}{A tibble of single-point mutants generated using `generate_mutants`}
}
\value{
a response matrix of the form \eqn{R_{ij}} (response site is \code{i}, mutated site is \code{j})
}
\description{
Calculate site-dependent structural response matrices
}
\details{
`calculate_df2ij_mrs()` calculates the force matrix \code{df2(i, j)} averaged over mutations at \code{j}

`calculate_de2ij_mrs()` calculates the energy-difference matrix \code{de2(i, j)} averaged over mutations at \code{j}

`calculate_dr2ij_mrs()` calculates the structural difference matrix \code{dr2(i, j)} averaged over mutations at \code{j}

`calculate_dvsij_mrs()` calculates the stress-energy matrix \code{dvs(i, j)} averaged over mutations at \code{j}

`calculate_dmsfij_mrs()` calculates the change of msf of site i due to mutations at j, averaged over mutations at \code{j}

`calculate_dhij_mrs()` calculates the change in entropy of site i due to mutations at j averaged over mutations at \code{j}

`calculate_rwsipij_mrs()` calculates rwsip between mutant and wt site i distributions due to mutations at j, averaged over mutations at j

`calculate_dbhatij_mrs()` calculates dbhat distance between wt and mut site i distributions, averaged over mutations at j.
}
\keyword{internal}
